+++ x/epochs/genesis_test.go
@@ -1,6 +1,7 @@
 package epochs_test
 
 import (
+	"github.com/EscanBE/evermint/v12/constants"
 	"testing"
 	"time"
 
@@ -19,7 +20,8 @@ func TestEpochsExportGenesis(t *testing.T) {
 	feemarketGenesis.Params.EnableHeight = 1
 	feemarketGenesis.Params.NoBaseFee = false
 
-	app := chainapp.Setup(false, feemarketGenesis)
+	chainID := constants.TestnetFullChainId
+	app := chainapp.Setup(false, feemarketGenesis, chainID)
 	ctx := app.BaseApp.NewContext(false, tmproto.Header{})
 
 	chainStartTime := ctx.BlockTime()
@@ -50,7 +52,8 @@ func TestEpochsInitGenesis(t *testing.T) {
 	feemarketGenesis.Params.EnableHeight = 1
 	feemarketGenesis.Params.NoBaseFee = false
 
-	app := chainapp.Setup(false, feemarketGenesis)
+	chainID := constants.TestnetFullChainId
+	app := chainapp.Setup(false, feemarketGenesis, chainID)
 	ctx := app.BaseApp.NewContext(false, tmproto.Header{})
 
 	// On init genesis, default epochs information is set